requirements.txt
# -------------------------------
# Core Python & Utilities
# -------------------------------
numpy
tqdm
Pillow

# -------------------------------
# Computer Vision
# -------------------------------
opencv-python

# -------------------------------
# Video & Audio Processing
# -------------------------------
moviepy

# -------------------------------
# Deep Learning (PyTorch)
# -------------------------------
torch
torchvision
torchaudio

# -------------------------------
# Face Detection
# -------------------------------
facenet-pytorch

# -------------------------------
# Optional / Stability
# -------------------------------
scipy
